Move test cases to use SyntaxException.
